uuid map metadata comparison endpoint
