+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - [#1126](https://github.com/LayerManager/layman/issues/1126) Endpoint [Workspace Layer Chunk](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#workspace-layer-chunk) (with methods GET, POST) was removed and replaced with endpoint [Layer Chunk](doc/rest.md#layer-chunk) to use UUID-based URL `/rest/layers/{uuid}/chunk` instead of workspace&name-based URL.
 - [#1126](https://github.com/LayerManager/layman/issues/1126) Endpoint [Workspace Maps](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#workspace-maps) (with methods GET, POST, DELETE) was unified into endpoint [Maps](doc/rest.md#get-maps); `workspace` is now supplied via request query/body parameter instead of URL path.
 - [#1126](https://github.com/LayerManager/layman/issues/1126) Endpoint [GET Workspace Layer Metadata Comparison](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-layer-metadata-comparison) was removed and replaced with endpoint [GET Layer Metadata Comparison](doc/rest.md#get-layer-metadata-comparison) to use UUID-based URL `/rest/layers/{uuid}/metadata-comparison` instead of workspace&name-based URL.
+- [#1126](https://github.com/LayerManager/layman/issues/1126) Endpoint [GET Workspace Map Metadata Comparison](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-map-metadata-comparison) was removed and replaced with endpoint [GET Map Metadata Comparison](doc/rest.md#get-map-metadata-comparison) to use UUID-based URL `/rest/maps/{uuid}/metadata-comparison` instead of workspace&name-based URL.
 
 ## v2.1.0
  2025-05-02
@@ -190,7 +191,7 @@
   - [GET](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-layer)/[PATCH](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#patch-workspace-layer)/[DELETE](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#delete-workspace-layer) Workspace Layer
   - GET Workspace Layer [Thumbnail](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-layer-thumbnail)/[Style](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-layer-style)/[Metadata Comparison](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-layer-metadata-comparison)/[Chunk](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-layer-chunk)
   - [GET](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-map)/[PATCH](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#patch-workspace-map)/[DELETE](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#delete-workspace-map) Workspace Map
-  - GET Workspace Map [File](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-map-file)/[Thumbnail](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-map-thumbnail)/[Metadata Comparison](doc/rest.md#get-workspace-map-metadata-comparison)
+  - GET Workspace Map [File](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-map-file)/[Thumbnail](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-map-thumbnail)/[Metadata Comparison](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-map-metadata-comparison)
   - GET Workspace [Layers](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-layers)/[Maps](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-maps)
   - GET [Layers](doc/rest.md#get-layers)/[Maps](doc/rest.md#get-maps)/[Publications](doc/rest.md#get-publications)
   - DELETE Workspace [Layers](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#delete-workspace-layers)/[Maps](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#delete-workspace-maps)
@@ -289,7 +290,7 @@
   - [POST](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#post-workspace-layers) and [DELETE](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#delete-workspace-layers) Workspace Layers
   - [POST](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#post-workspace-maps) and [DELETE](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#delete-workspace-maps) Workspace Maps
   - requests to [WMS](doc/endpoints.md#web-map-service) and [WFS](doc/endpoints.md#web-feature-service) endpoints
-- [#868](https://github.com/LayerManager/layman/issues/868) Responses to [GET Workspace Layer Metadata Comparison](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-layer-metadata-comparison) and [GET Workspace Map Metadata Comparison](doc/rest.md#get-workspace-map-metadata-comparison) do not respect [HTTP X-Forwarded headers](doc/client-proxy.md#x-forwarded-http-headers) of the request intentionally, in order to keep URLs in canonical form.
+- [#868](https://github.com/LayerManager/layman/issues/868) Responses to [GET Workspace Layer Metadata Comparison](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-layer-metadata-comparison) and [GET Workspace Map Metadata Comparison](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-map-metadata-comparison) do not respect [HTTP X-Forwarded headers](doc/client-proxy.md#x-forwarded-http-headers) of the request intentionally, in order to keep URLs in canonical form.
 - [#868](https://github.com/LayerManager/layman/issues/868) Relations between map and [internal layers](doc/models.md#internal-map-layer) are updated in `map_layer` table when calling [POST Workspace Maps](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#post-workspace-maps), [PATCH Workspace Map](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#patch-workspace-map), [DELETE Workspace Map](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#delete-workspace-map), and [DELETE Workspace Maps](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#delete-workspace-maps).
 - [#901](https://github.com/LayerManager/layman/issues/901) Endpoints [POST](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#post-workspace-maps) and [PATCH](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#patch-workspace-map) accept map compositions in version `3.x.x`.
 - [#927](https://github.com/LayerManager/layman/issues/927) SLD styles are internally published to GeoServer with [`raw`](https://docs-archive.geoserver.org/2.21.x/en/user/rest/api/styles.html#raw) parameter set to `True`.
@@ -1094,7 +1095,7 @@ There is a critical bug in this release, posting new layer breaks Layman: https:
 ### Changes
 - Both [PATCH Workspace Layer](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#patch-workspace-layer) and [PATCH Workspace Map](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#patch-workspace-map) automatically update also CSW metadata records.
 - Use absolute URLs in responses of Layer and Map endpoints
-- Add [GET Workspace Layer Metadata Comparison](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-layer-metadata-comparison) and [GET Workspace Map Metadata Comparison](doc/rest.md#get-workspace-map-metadata-comparison) endpoints. 
+- Add [GET Workspace Layer Metadata Comparison](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-layer-metadata-comparison) and [GET Workspace Map Metadata Comparison](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-map-metadata-comparison) endpoints.
 - Add [`revision_date`](doc/metadata.md#revision_date) metadata property
 - Add `metadata.comparison_url` to [GET Workspace Layer](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-layer) and [GET Workspace Map](https://github.com/LayerManager/layman/blob/v2.1.0/doc/rest.md#get-workspace-map) responses.
 - Upgrade Layman test client to version 1.1.1
