+++ test_tools/process_client.py
@@ -66,7 +66,7 @@ PUBLICATION_TYPES_DEF = {MAP_TYPE: PublicationTypeDef('mapname',
                                                       'rest_maps.delete',
                                                       map_keys_to_check,
                                                       'sample/layman.map/small_map.json',
-                                                      'rest_workspace_map_metadata_comparison.get',
+                                                      'rest_map_metadata_comparison.get',
                                                       None,
                                                       ),
                          LAYER_TYPE: PublicationTypeDef('layername',
@@ -492,13 +492,8 @@ def get_workspace_publication_metadata_comparison(publication_type, workspace, n
 
     publication_type_def = PUBLICATION_TYPES_DEF[publication_type]
     with app.app_context():
-        if publication_type == LAYER_TYPE:
-            uuid = layman_util.get_publication_uuid(workspace, publication_type, name)
-            r_url = url_for(publication_type_def.get_workspace_metadata_comparison_url, uuid=uuid)
-        else:
-            r_url = url_for(publication_type_def.get_workspace_metadata_comparison_url,
-                            **{publication_type_def.url_param_name: name},
-                            workspace=workspace)
+        uuid = layman_util.get_publication_uuid(workspace, publication_type, name)
+        r_url = url_for(publication_type_def.get_workspace_metadata_comparison_url, uuid=uuid)
     response = requests.get(r_url, headers=headers, timeout=HTTP_TIMEOUT)
     raise_layman_error(response)
     return response.json()
